+++ iphone/Classes/AccelerometerModule.m
@@ -61,7 +61,7 @@
     }
 }
 
--(void)_listenerAdded:(NSString *)type count:(int)count
+-(void)_listenerAdded:(NSString *)type count:(NSInteger)count
 {
     if ( (count == 1) && [type isEqualToString:@"update"] ) {
         [self startGeneratingEvents];
@@ -69,7 +69,7 @@
 	
 }
 
--(void)_listenerRemoved:(NSString *)type count:(int)count
+-(void)_listenerRemoved:(NSString *)type count:(NSInteger)count
 {
     if ( (count == 0) && [type isEqualToString:@"update"] ) {
         [self stopGeneratingEvents];
