+++ iphone/Classes/TiDOMNamedNodeMapProxy.m
@@ -103,7 +103,7 @@
 -(id)item:(id)args
 {
 	ENSURE_SINGLE_ARG(args,NSObject);
-	int index = [TiUtils intValue:args];
+	NSInteger index = [TiUtils intValue:args];
 
     if( ([[(GDataXMLElement*)[element node] attributes] count] > index) && (index >= 0) )
     {
